+++ src/php/Compiler/Domain/Lexer/Token.php
@@ -12,35 +12,37 @@ final readonly class Token
 
     public const int T_NEWLINE = 3;
 
-    public const int T_COMMENT = 4;
+    public const int T_COMMENT_MACRO = 4;
 
-    public const int T_UNQUOTE_SPLICING = 5;
+    public const int T_COMMENT = 5;
 
-    public const int T_OPEN_PARENTHESIS = 6;
+    public const int T_UNQUOTE_SPLICING = 6;
 
-    public const int T_CLOSE_PARENTHESIS = 7;
+    public const int T_OPEN_PARENTHESIS = 7;
 
-    public const int T_OPEN_BRACKET = 8;
+    public const int T_CLOSE_PARENTHESIS = 8;
 
-    public const int T_CLOSE_BRACKET = 9;
+    public const int T_OPEN_BRACKET = 9;
 
-    public const int T_OPEN_BRACE = 10;
+    public const int T_CLOSE_BRACKET = 10;
 
-    public const int T_CLOSE_BRACE = 11;
+    public const int T_OPEN_BRACE = 11;
 
-    public const int T_QUOTE = 12;
+    public const int T_CLOSE_BRACE = 12;
 
-    public const int T_UNQUOTE = 13;
+    public const int T_QUOTE = 13;
 
-    public const int T_QUASIQUOTE = 14;
+    public const int T_UNQUOTE = 14;
 
-    public const int T_CARET = 15;
+    public const int T_QUASIQUOTE = 15;
 
-    public const int T_FN = 16;
+    public const int T_CARET = 16;
 
-    public const int T_STRING = 17;
+    public const int T_FN = 17;
 
-    public const int T_ATOM = 18;
+    public const int T_STRING = 18;
+
+    public const int T_ATOM = 19;
 
     public const int T_EOF = 100;