+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Fix autoload phel classes from phar
 - Fix vector and hash map literals in try
 - Fix REPL symbol resolution issue loading current dir
+- Add inline comment with macro `#_`
 
 ## [0.20.0](https://github.com/phel-lang/phel-lang/compare/v0.19.1...v0.20.0) - 2025-08-25
 
@@ -43,6 +44,7 @@ All notable changes to this project will be documented in this file.
 - Add new macros:
     - `some->`
     - `some->>`
+- Add `#_` reader macro for inline comments
 - Extend `take` to handle PHP Traversable inputs
     - enabling safe slicing of generators and other iterables for infinite list scenarios
 - Enhance `php/->` for nested calls